the making of a game...
Don't look back! is the title
I'm betting on the player's curiosity to actually look back
which will start the game and reveal the outlaws

Player will shoot outlaws with gaze. As you shoot them or turn your back
to them, some spring back. There should be an order to shoot them in
that prevent them from spawning back and win the game.